myProjects/hugoApp/requirements.txt
autopep8==1.5.7
flake8==3.9.2
fpdf2==2.4.2
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
Pillow==8.3.1
pycodestyle==2.7.0
pyflakes==2.3.1
toml==0.10.2
typing-extensions==3.10.0.0
